Remove deprecated references
