+++ text2vec/models/__init__.py
@@ -1,8 +1,7 @@
 from text2vec.models.transformer import TransformerEncoder, TransformerDecoder
 from text2vec.models.sequential import RecurrentEncoder, RecurrentDecoder
 
-from text2vec.models.components.feeder import TextInput, Tokenizer
-from text2vec.models.components.text_inputs import Embed, TokenEmbed
+from text2vec.models.components.text_inputs import Embed, TokenEmbed, Tokenizer
 from text2vec.models.components.attention import BahdanauAttention, MultiHeadAttention
 from text2vec.models.components.feed_forward import PositionWiseFFN
 from text2vec.models.components.recurrent import BidirectionalLSTM
